Move slack to models
